pyproject.toml
[tool.poetry]
name = "crispradium"
version = "1.0.0"
description = "A web application that helps researchers design optimal CRISPR guide RNA sequences. It analyzes input DNA, identifies potential target sites, and scores the guide RNAs based on factors like GC content, secondary structure, and off-target potential."
authors = ["Bjorn99 <67769176+Bjorn99@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.3"
Flask = "^3.0.3"
biopython = "^1.85"
python-requirements = "^3.15.6"
requests = "^2.32.3"
flake8 = "^7.1.1"
black = "^24.10.0"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
viennarna = "^2.7.0"
rna = "^0.13.2"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
